+++ src/main/java/com/abc/services/FollowService.java
@@ -1,13 +1,12 @@
 package com.abc.services;
 
 import java.util.List;
-
 import com.abc.entities.User;
 
 public interface FollowService {
-    List<User> getUserFollower(int id);
-    List<User> getUserFollowed(int id);
-    List<User> getSuggestFollow(int id);
-    void followUser(int followingUserId, int followedUserId);
-    void unfollowUser(int followingUserId, int followedUserId);
+    List<User> getUserFollower(Long id);
+    List<User> getUserFollowed(Long id);
+    List<User> getSuggestFollow(Long id);
+    void followUser(Long followingUserId, Long followedUserId);
+    void unfollowUser(Long followingUserId, Long followedUserId);
 }
