+++ src/main/java/com/abc/services/PostService.java
@@ -1,11 +1,10 @@
 package com.abc.services;
 
 import java.util.List;
-
 import com.abc.entities.Post;
 
 public interface PostService {
-    List<Post> getAllPost(int id);
-    List<Post> getPostById(int id);
+    List<Post> getAllPost(Long id);
+    List<Post> getPostById(Long id);
     boolean createdPost(Post post);
 }
