+++ src/main/java/com/abc/services/PostServiceImpl.java
@@ -1,26 +1,24 @@
 package com.abc.services;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.abc.dao.PostDAO;
 import com.abc.entities.Post;
 
-@Service // Spring sẽ quản lý bean này
+@Service
 public class PostServiceImpl implements PostService {
 
-	@Autowired
+    @Autowired
     private PostDAO postDAO;
 
     @Override
-    public List<Post> getAllPost(int id) {
+    public List<Post> getAllPost(Long id) {
         return postDAO.getALLPost(id);
     }
 
     @Override
-    public List<Post> getPostById(int id) {
+    public List<Post> getPostById(Long id) {
         return postDAO.getPostById(id);
     }
